refactor: relocate the assistant into the shared assistantview (#1616)
