--- app/components/Assistant/index.ts
@@ -1,4 +1,0 @@
-export { ChatMessage } from "./ChatMessage/chatMessage";
-export { ChatPanel } from "./ChatPanel/chatPanel";
-export { SchemaPanel } from "./SchemaPanel/schemaPanel";
-export { SuggestionChips } from "./SuggestionChips/suggestionChips";
