+++ packages/shared/views/AssistantView/assistantView.styles.ts
