+++ packages/shared/views/AssistantView/components/ChatPanel/chatPanel.styles.ts
