+++ packages/shared/views/AssistantView/components/ChatPanel/chatPanel.tsx
@@ -1,5 +1,3 @@
-import { ChatMessage } from "@/components/Assistant/ChatMessage/chatMessage";
-import { SuggestionChips } from "@/components/Assistant/SuggestionChips/suggestionChips";
 import {
   Alert,
   Box,
@@ -10,6 +8,8 @@ import {
 } from "@mui/material";
 import { useAuth } from "@repo/shared/providers/authentication/provider";
 import type { SuggestionChip } from "@repo/shared/services/api-client/types";
+import { ChatMessage } from "@repo/shared/views/AssistantView/components/ChatMessage/chatMessage";
+import { SuggestionChips } from "@repo/shared/views/AssistantView/components/SuggestionChips/suggestionChips";
 import { type JSX, useEffect, useRef, useState } from "react";
 import { ChatContainer, InputRow, MessagesContainer } from "./chatPanel.styles";
