+++ packages/shared/views/AssistantView/components/SchemaPanel/schemaPanel.styles.ts
