+++ packages/shared/views/AssistantView/components/SchemaPanel/schemaPanel.tsx
