+++ packages/shared/views/AssistantView/components/SchemaPanel/utils.ts
